refactor(cli): extract shared `runSdkDocker` helper
`callDockerBuild` and `testService` were near-identical, differing only
in `sdk.build` vs `sdk.test` and the failure-message verb. Move the
shared logic (plus `DockerError`) into `bin/cli/src/utils/run-sdk-docker.ts`
and reduce each command's docker entry point to a one-line delegation.
